Python_Bootcamp._Day_06-0/src/ex01/requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
names==0.3.0
pydantic-1.10.2



